Start attempting to clean up circular dependencies
Try to untangle the mess of client-side circular dependencies by first ensuring
Common.js doesn't depend on anything:

1. Move server communication to its own file
2. Move FetchError into its own file
3. Move error handling methods into their own file
4. Trigger ServerPaused overlay with an event instead of a direct call to avoid
   depending on Overlay and everything that brings in.

This reduces the number of circular dependencies from 53 to 41 according to dpdm
