+++ Client/Script/Common.js
@@ -1,27 +1,3 @@
-import { BulkMarkerResolveType } from '../../Shared/PlexTypes.js';
-import { ContextualLog } from '../../Shared/ConsoleLog.js';
-import { MarkerEnum } from '../../Shared/MarkerType.js';
-
-import { animate } from './AnimationHelpers.js';
-import Overlay from './Overlay.js';
-import ServerPausedOverlay from './ServerPausedOverlay.js';
-
-/** @typedef {!import('../../Shared/PlexTypes').BulkDeleteResult} BulkDeleteResult */
-/** @typedef {!import('../../Shared/PlexTypes').BulkRestoreResponse} BulkRestoreResponse */
-/** @typedef {!import('../../Shared/PlexTypes').ChapterMap} ChapterMap */
-/** @typedef {!import('../../Shared/PlexTypes').CustomBulkAddMap} CustomBulkAddMap */
-/** @typedef {!import('../../Shared/PlexTypes').ExtendedQueryInfo} ExtendedQueryInfo */
-/** @typedef {!import('../../Shared/PlexTypes').PurgeSection} PurgeSection */
-/** @typedef {!import('../../Shared/PlexTypes').SerializedBulkAddResult} SerializedBulkAddResult */
-/** @typedef {!import('../../Shared/PlexTypes').SerializedEpisodeData} SerializedEpisodeData */
-/** @typedef {!import('../../Shared/PlexTypes').SerializedMarkerData} SerializedMarkerData */
-/** @typedef {!import('../../Shared/PlexTypes').SerializedMovieData} SerializedMovieData */
-/** @typedef {!import('../../Shared/PlexTypes').SerializedSeasonData} SerializedSeasonData */
-/** @typedef {!import('../../Shared/PlexTypes').SerializedShowData} SerializedShowData */
-/** @typedef {!import('../../Shared/PlexTypes').ShiftResult} ShiftResult */
-/** @typedef {!import('../../Shared/MarkerBreakdown').MarkerBreakdownMap} MarkerBreakdownMap */
-
-const Log = new ContextualLog('ClientCommon');
 
 /**
  * Removes all children from the given element.
@@ -33,308 +9,6 @@ function clearEle(ele) {
     }
 }
 
-/**
- * Custom error class used to distinguish between errors
- * surfaced by an API call and all others. */
-class FetchError extends Error {
-    /**
-     * @param {string} message
-     * @param {string} stack */
-    constructor(message, stack) {
-        super(message);
-        if (stack) {
-            this.stack = stack;
-        }
-    }
-}
-
-/* eslint-disable max-len */
-/**
- * Map of all available server endpoints.
- * Keep in sync with ServerCommands.#commandMap */
-const ServerCommand = {
-    /**
-     * Add a marker to the Plex database
-     * @param {string} type
-     * @param {number} metadataId
-     * @param {number} start
-     * @param {number} end
-     * @param {number} [final=0]
-     * @returns {Promise<SerializedMarkerData>} */
-    add : (type, metadataId, start, end, final=0) => jsonRequest('add', { metadataId, start, end, type, final }),
-
-    /**
-     * Edit an existing marker with the given id.
-     * @param {string} type marker type
-     * @param {number} id
-     * @param {number} start
-     * @param {number} end
-     * @param {number} [final=0]
-     * @returns {Promise<SerializedMarkerData>} */
-    edit : (type, id, start, end, final=0) => jsonRequest('edit', { id, start, end, type, final }),
-
-    /**
-     * Delete the marker with the given id.
-     * @param {number} id
-     * @returns {Promise<SerializedMarkerData>} */
-    delete : (id) => jsonRequest('delete', { id }),
-
-    /**
-     * Retrieve all markers under the given metadata id that may be affected by a shift operation.
-     * @param {number} id
-     * @param {number} applyTo
-     * @returns {Promise<ShiftResult>} */
-    checkShift : (id, applyTo) => jsonRequest('check_shift', { id, applyTo }),
-
-    /**
-     * Shift all markers under the given metadata by the given shift, unless they're in the ignored list
-     * @param {number} id The metadata id of the item to shift markers for.
-     * @param {number} startShift The number of milliseconds to shift marker starts (positive or negative).
-     * @param {number} endShift The number of milliseconds to shift marker ends (positive or negative).
-     * @param {number} applyTo The marker type(s) to apply the shift to.
-     * @param {boolean} force False to abort if there are episodes with multiple markers, true to shift all markers regardless.
-     * @param {number[]?} [ignored=[]] Array of marker ids to ignore when shifting.
-     * @returns {Promise<ShiftResult>} */
-    shift : (id, startShift, endShift, applyTo, force, ignored=[]) => jsonRequest('shift', { id : id, startShift : startShift, endShift : endShift, applyTo : applyTo, force : force ? 1 : 0, ignored : ignored.join(',') }),
-
-    /**
-     * Query for all markers that would be deleted for the given metadata id.
-     * @param {number} id
-     * @returns {Promise<BulkDeleteResult>} */
-    checkBulkDelete : (id) => jsonRequest('bulk_delete', { id : id, dryRun : 1, applyTo : MarkerEnum.All, ignored : [] }),
-
-    /**
-     * Delete all markers associated with the given media item, except those specified in `ignored`.
-     * @param {number} id
-     * @param {number} applyTo The marker type(s) to apply the delete to.
-     * @param {number[]} [ignored =[]] List of marker ids to not delete.
-     * @returns {Promise<BulkDeleteResult>} */
-    bulkDelete : (id, applyTo, ignored=[]) => jsonRequest('bulk_delete', { id : id, dryRun : 0, applyTo : applyTo, ignored : ignored.join(',') }),
-
-    /**
-     * Retrieve episode and marker information relevant to a bulk_add operation.
-     * @param {number} id Show/Season metadata id.
-     * @returns {Promise<SerializedBulkAddResult>} */
-    checkBulkAdd : (id) => jsonRequest('bulk_add', { id : id, start : 0, end : 0, resolveType : BulkMarkerResolveType.DryRun, ignored : '', type : 'intro' }),
-
-    /**
-     * Bulk adds a marker to the given metadata id.
-     * @param {string} markerType The type of marker (intro/credits)
-     * @param {number} id Show/Season metadata id.
-     * @param {number} start Start time of the marker, in milliseconds.
-     * @param {number} end End time of the marker, in milliseconds.
-     * @param {number} resolveType The BulkMarkerResolveType.
-     * @param {number[]?} ignored The list of episode ids to ignore adding markers to.
-     * @returns {Promise<SerializedBulkAddResult>} */
-    bulkAdd : (markerType, id, start, end, resolveType, ignored=[]) => jsonRequest('bulk_add', { id : id, start : start, end : end, type : markerType, resolveType : resolveType, ignored : ignored.join(',') }),
-
-    /**
-     * Bulk adds multiple markers with custom timestamps.
-     * @param {number} markerType The type of marker (intro/credits)
-     * @param {number} id The Show/Season metadata id.
-     * @param {number} resolveType The BulkMarkerResolveType.
-     * @param {CustomBulkAddMap} newMarkerData The new markers to add.
-     * @returns {Promise<SerializedBulkAddResult>} */
-    bulkAddCustom : (markerType, id, resolveType, newMarkerData) => jsonBodyRequest('add_custom', { type : markerType, id : id, resolveType : resolveType, markers : JSON.stringify(newMarkerData) }),
-
-    /**
-     * Retrieve markers for all episodes ids in `keys`.
-     * @param {number[]} keys The list of episode ids to grab the markers of.
-     * @returns {Promise<{[metadataId: number]: SerializedMarkerData[]}>} */
-    query : (keys) => jsonRequest('query', { keys : keys.join(',') }),
-
-    /**
-     * Retrieve all Movie/TV library sections in the database.
-     * @returns {Promise<{id: number, type : number, name: string}[]} */
-    getSections : () => jsonRequest('get_sections'),
-
-    /**
-     * Retrieve all shows in the given section.
-     * @param {number} id
-     * @returns {Promise<SerializedShowData[]|SerializedMovieData[]>} */
-    getSection : (id) => jsonRequest('get_section', { id }),
-
-    /**
-     * Retrieve all seasons in the given show.
-     * @param {number} id
-     * @returns {Promise<SerializedSeasonData[]>} */
-    getSeasons : (id) => jsonRequest('get_seasons', { id }),
-
-    /**
-     * Retrieve all episodes in the given season.
-     * @param {number} id
-     * @returns {Promise<SerializedEpisodeData>} */
-    getEpisodes : (id) => jsonRequest('get_episodes', { id }),
-
-    /**
-     * Return whether the given metadata item has thumbnails associated with it.
-     * Only valid for episode/movie metadata ids.
-     * @param {number} id
-     * @returns {Promise<{hasThumbnails: boolean}>} */
-    checkForThumbnails : (id) => jsonRequest('check_thumbs', { id }),
-
-    /**
-     * Retrieve marker breakdown stats for the given section.
-     * @param {number} id
-     * @returns {Promise<{[episodesWithNMarkers: number]: number}>} */
-    getMarkerStats : (id) => jsonRequest('get_stats', { id }),
-
-    /**
-     * Retrieve the marker breakdown stats for a single show or movie.
-     * @param {number} id
-     * @param {boolean} includeSeasons True to include season data, false to leave it out (or if it's for a movie).
-     * @returns {Promise<{mainData: MarkerBreakdownMap, seasonData?: { [seasonId: number]: MarkerBreakdownMap }}>} */
-    getBreakdown : (id, includeSeasons) => jsonRequest('get_breakdown', { id : id, includeSeasons : includeSeasons ? 1 : 0 }),
-
-
-    /**
-     * Retrieve the configuration settings relevant to the client application.
-     * @returns {Promise<{userThumbnails: boolean, extendedMarkerStats: boolean, version: string }>} */
-    getConfig : () => jsonRequest('get_config'),
-
-    /**
-     * Set server-side log settings.
-     * @param {number} level The log level.
-     * @param {number} dark 1 to color messages for a dark console (if supported), 0 for light mode.
-     * @param {number} trace 1 to log stack traces, 0 otherwise */
-    logSettings : (level, dark, trace) => jsonRequest('log_settings', { level, dark, trace }),
-
-
-    /**
-     * Check for markers that should exist for the given metadata id, but aren't in the Plex database.
-     * @param {number} id The show/season/episode id.
-     * @returns {Promise<SerializedMarkerData[]>} */
-    purgeCheck : (id) => jsonRequest('purge_check', { id }),
-
-    /**
-     * Find all purges in the given library section.
-     * @param {number} sectionId
-     * @returns {Promise<PurgeSection>} */
-    allPurges : (sectionId) => jsonRequest('all_purges', { sectionId }),
-
-    /**
-     * Restore the given purged markers associated with the given section.
-     * @param {number[]} markerIds Array of purged marker ids to restore.
-     * @param {number} sectionId
-     * @param {number} resolveType
-     * @returns {Promise<BulkRestoreResponse>} */
-    restorePurge : (markerIds, sectionId, resolveType) => jsonRequest('restore_purge', { markerIds : markerIds.join(','), sectionId : sectionId, resolveType : resolveType }),
-
-    /**
-     * Ignore the given purged markers associated with the given section.
-     * @param {number[]} markerIds
-     * @param {number} sectionId
-     * @returns {Promise<void>} */
-    ignorePurge : (markerIds, sectionId) => jsonRequest('ignore_purge', { markerIds, sectionId }),
-
-    /**
-     * Irreversibly delete all markers of the given type from the given section.
-     * @param {number} sectionId
-     * @param {number} deleteType
-     * @returns {Promise<{ deleted : number, backupDeleted : number, cacheDeleted : number }>} */
-    sectionDelete : (sectionId, deleteType) => jsonRequest('nuke_section', { sectionId, deleteType }),
-
-    /**
-     * Shutdown Marker Editor.
-     * @returns {Promise<void>} */
-    shutdown : () => jsonRequest('shutdown'),
-    /**
-     * Restart Marker Editor.
-     * @returns {Promise<void>} */
-    restart : () => jsonRequest('restart'),
-    /**
-     * Reload all markers from the database. Like restart, but doesn't also restart the HTTP server.
-     * @returns {Promise<void>} */
-    reload : () => jsonRequest('reload'),
-    /**
-     * Suspend Marker Editor.
-     * @returns {Promise<void>} */
-    suspend : () => jsonRequest('suspend'),
-    /**
-     * Resume a suspended Marker Editor.
-     * @returns {Promise<void>} */
-    resume : () => jsonRequest('resume'),
-
-    /**
-     * Upload a database file and import the markers present into the given section.
-     * @param {Object} database
-     * @param {number} sectionId
-     * @param {number} resolveType */
-    importDatabase : (database, sectionId, resolveType) => jsonBodyRequest('import_db', { database, sectionId, resolveType }),
-
-    /**
-     * Retrieve chapter data (if any) for the given media item (supports shows, seasons, episodes, and movies).
-     * @param {number} metadataId
-     * @returns {Promise<ChapterMap>} */
-    getChapters : (metadataId) => jsonRequest('get_chapters', { id : metadataId }),
-
-    /**
-     * Retrieve all information relevant for marker table creation for a given movie/episode id.
-     * @param {number} metadataId
-     * @returns {Promise<ExtendedQueryInfo} */
-    extendedQuery : (metadataId) => jsonRequest('query_full', { id : metadataId }),
-};
-/* eslint-enable */
-
-/**
- * Generic method to make a request to the given endpoint that expects a JSON response.
- * @param {string} endpoint The URL to query.
- * @param {{[parameter: string]: any}} parameters URL parameters. */
-function jsonRequest(endpoint, parameters={}) {
-    const url = new URL(endpoint, window.location.href);
-    for (const [key, value] of Object.entries(parameters)) {
-        url.searchParams.append(key, value);
-    }
-
-    return jsonPostCore(url);
-}
-
-/**
- * Similar to jsonRequest, but expects blob data and attaches parameters to the body instead of URL parameters.
- * @param {string} endpoint
- * @param {Object} parameters */
-function jsonBodyRequest(endpoint, parameters={}) {
-    const url = new URL(endpoint, window.location.href);
-    const data = new FormData();
-    for (const [key, value] of Object.entries(parameters)) {
-        data.append(key, value);
-    }
-
-    return jsonPostCore(url, data);
-}
-
-/**
- * Core method that makes a request to the server, expecting JSON in return.
- * @param {URL} url The fully built URL endpoint
- * @param {FormData} body The message body, if any. */
-async function jsonPostCore(url, body=null) {
-    const init = { method : 'POST', headers : { accept : 'application/json' } };
-    if (body) {
-        init.body = body;
-    }
-
-    try {
-        const response = await (await fetch(url, init)).json();
-        Log.verbose(response, `Response from ${url}`);
-        if (!response || response.Error) {
-
-            // Global check to see if we failed because the server is suspended.
-            // If so, show the undismissible 'Server Paused' overlay.
-            if (response.Error && response.Error === 'Server is suspended') {
-                Log.info('Action was not completed because the server is suspended.');
-                ServerPausedOverlay.Show();
-                // Return unfulfillable Promise. Gross, but since the user can't do anything anyway, we don't really care.
-                return new Promise((_resolve, _reject) => {});
-            }
-
-            throw new FetchError(response ? response.Error : `Request to ${url} failed`);
-        }
-
-        return response;
-    } catch (err) {
-        throw new FetchError(err.message, err.stack);
-    }
-}
 
 /**
  * Custom jQuery-like selector method.
@@ -440,68 +114,6 @@ function appendChildren(parent, ...elements) {
     return parent;
 }
 
-
-/**
- * Return an error string from the given error.
- * In almost all cases, `error` will be either a JSON object with a single `Error` field,
- * or an exception of type {@link Error}. Handle both of those cases, otherwise return a
- * generic error message.
- *
- * NOTE: It's expected that all API requests call this on failure, as it's the main console
- *       logging method.
- * @param {string|Error} error
- * @returns {string} */
-function errorMessage(error) {
-    if (error.Error) {
-        Log.error(error);
-        return error.Error;
-    }
-
-    if (error instanceof Error) {
-        Log.error(error.message);
-        Log.error(error.stack ? error.stack : '(Unknown stack)');
-
-        if (error instanceof TypeError && error.message === 'Failed to fetch') {
-            // Special handling of what's likely a server-side exit.
-            return error.toString() + '<br><br>The server may have exited unexpectedly, please check the console.';
-        }
-
-        return error.toString();
-    }
-
-    if (typeof error === 'string') {
-        return error;
-    }
-
-    return 'I don\'t know what went wrong, sorry :(';
-}
-
-/**
- * Displays an error message in the top-left of the screen for a couple seconds.
- * @param {string|HTMLElement} message */
-function errorToast(message, duration=2500) {
-    const msg = buildNode('div', { class : 'errorToast' }, message);
-    const container = $('#toastContainer');
-    container.appendChild(msg);
-
-    // Hack based on known css padding/border heights to avoid getComputedStyle.
-    const height = (msg.getBoundingClientRect().height - 32) + 'px';
-    msg.style.height = height;
-
-    return animate(msg,
-        [
-            { opacity : 0 },
-            { opacity : 1, offset : 0.2 },
-            { opacity : 1, offset : 0.8 },
-            { height : height, overflow : 'hidden', padding : '15px', offset : 0.95 },
-            { opacity : 0, height : '0px', overflow : 'hidden', padding : '0px 15px 0px 15px', offset : 1 },
-        ],
-        { duration },
-        () => {
-            container.removeChild(msg);
-        }
-    );
-}
 
 /**
  * Return 'n text' if n is 1, otherwise 'n texts'.
@@ -751,26 +363,6 @@ function clickOnEnterCallback(e) {
 }
 
 /**
- * Displays an overlay for the given error
- * @param {string} message
- * @param {Error|string} err
- * @param {() => void} [onDismiss=Overlay.dismiss] */
-function errorResponseOverlay(message, err, onDismiss=Overlay.dismiss) {
-    const errType = err instanceof FetchError ? 'Server Message' : 'Error';
-    Overlay.show(
-        appendChildren(
-            buildNode('div'),
-            document.createTextNode(message),
-            buildNode('br'),
-            buildNode('br'),
-            document.createTextNode(errType + ':'),
-            buildNode('br'),
-            document.createTextNode(errorMessage(err))),
-        'OK',
-        onDismiss);
-}
-
-/**
  * Waits for the given condition to be true, timing out after a specified number of milliseconds.
  * @param {() => boolean} condition The condition to test until it returns true.
  * @param {number} timeout Number of milliseconds before giving up.
@@ -796,15 +388,11 @@ export {
     buildNodeNS,
     clearEle,
     clickOnEnterCallback,
-    errorMessage,
-    errorToast,
-    errorResponseOverlay,
     msToHms,
     pad0,
     plural,
     realMs,
     roundDelta,
-    ServerCommand,
     timeInputShortcutHandler,
     timeToMs,
     waitFor,
