+++ Client/Script/CustomEvents.js
@@ -1,0 +1,8 @@
+/**
+ * List of Marker Editor's custom events. */
+export const CustomEvents = {
+    /** @readonly Triggered when the stickiness setting changes. */
+    StickySettingsChanged : 'stickySettingsChanged',
+    /** @readonly Triggered when we attempt to reach out to the server when it's paused. */
+    ServerPaused : 'serverPaused',
+};
