+++ Client/Script/ServerPausedOverlay.js
@@ -1,12 +1,17 @@
-import { errorResponseOverlay, ServerCommand } from './Common.js';
-
+import { CustomEvents } from './CustomEvents.js';
+import { errorResponseOverlay } from './ErrorHandling.js';
 import Overlay from './Overlay.js';
+import { ServerCommands } from './Commands.js';
 
 /**
  * Static class that is responsible for displaying an undismissible overlay letting the user know
  * that the server is suspended, giving them the option to resume it.
  */
 class ServerPausedOverlay {
+    static Setup() {
+        window.addEventListener(CustomEvents.ServerPaused, ServerPausedOverlay.Show);
+    }
+
     static Show() {
         Overlay.show(
             `Server is Paused. Press 'Resume' to reconnect to the Plex database.`,
@@ -22,7 +27,7 @@ class ServerPausedOverlay {
 async function onResume(_, button) {
     button.innerText = 'Resuming...';
     try {
-        await ServerCommand.resume();
+        await ServerCommands.resume();
         window.location.reload();
     } catch (err) {
         button.innerText = 'Resume';
